最新config

+++ cpct-service/src/main/java/com/zjtelcom/cpct/service/impl/channel/LabelServiceImpl.java
@@ -191,9 +191,11 @@ public class LabelServiceImpl extends BaseService implements LabelService {
         label.setStatusCd("1000");
         labelMapper.insert(label);
         Long labelId = label.getInjectionLabelId();
-        label.setInjectionLabelId(100000+label.getInjectionLabelId());
-        labelMapper.insert(label);
-        labelMapper.deleteByPrimaryKey(labelId);
+        if (labelId<10000){
+            label.setInjectionLabelId(100000+label.getInjectionLabelId());
+            labelMapper.insert(label);
+            labelMapper.deleteByPrimaryKey(labelId);
+        }
         insertLabelValue(label,addVO.getRightOperand());
 
         //redis更新标签库
@@ -210,7 +212,7 @@ public class LabelServiceImpl extends BaseService implements LabelService {
         result.put("resultCode",CODE_SUCCESS);
         result.put("resultMsg","添加成功");
 
-        if (value.equals("1")){
+        if (SystemParamsUtil.getSyncValue().equals("1")){
             new Thread(){
                 public void run(){
                     try {
@@ -271,7 +273,7 @@ public class LabelServiceImpl extends BaseService implements LabelService {
         result.put("resultCode",CODE_SUCCESS);
         result.put("resultMsg","编辑成功");
 
-        if (value.equals("1")){
+        if (SystemParamsUtil.getSyncValue().equals("1")){
             new Thread(){
                 public void run(){
                     try {
@@ -355,7 +357,7 @@ public class LabelServiceImpl extends BaseService implements LabelService {
         result.put("resultCode",CODE_SUCCESS);
         result.put("resultMsg","删除成功");
 
-        if (value.equals("1")){
+        if (SystemParamsUtil.getSyncValue().equals("1")){
             new Thread(){
                 public void run(){
                     try {
@@ -427,7 +429,7 @@ public class LabelServiceImpl extends BaseService implements LabelService {
         result.put("resultCode",CODE_SUCCESS);
         result.put("resultMsg","添加成功");
 
-        if (value.equals("1")){
+        if (SystemParamsUtil.getSyncValue().equals("1")){
             new Thread(){
                 public void run(){
                     try {
@@ -458,7 +460,7 @@ public class LabelServiceImpl extends BaseService implements LabelService {
         result.put("resultCode",CODE_SUCCESS);
         result.put("resultMsg","添加成功");
 
-        if (value.equals("1")){
+        if (SystemParamsUtil.getSyncValue().equals("1")){
             new Thread(){
                 public void run(){
                     try {
@@ -489,7 +491,7 @@ public class LabelServiceImpl extends BaseService implements LabelService {
             result.put("resultCode", CODE_SUCCESS);
             result.put("resultMsg", "删除成功");
 
-            if (value.equals("1")){
+            if (SystemParamsUtil.getSyncValue().equals("1")){
                 new Thread(){
                     public void run(){
                         try {
@@ -591,7 +593,7 @@ public class LabelServiceImpl extends BaseService implements LabelService {
         result.put("resultCode",CODE_SUCCESS);
         result.put("resultMsg","添加成功");
 
-        if (value.equals("1")){
+        if (SystemParamsUtil.getSyncValue().equals("1")){
             new Thread(){
                 public void run(){
                     try {
